Bug fix for Zernike sensitivity control.
Bug was fixed where Zernike sensitivities couldn't be suppressed if the size of the pupil array (mp.P1.compact.Narr and mp.P1.full.Narr) was defined differently from the default of being just large enough to contain the aperture. The array for the Zernike mode is now padded/cropped to the same size.