Code at Time of SPIE
This is the state of the code just before the 2019 SPIE conference.  The current base parameters should get close to the multicore fiber experiments we plan to run on HCST at Caltech.  The vortex and SPLC channels should be fully functional; the others have not yet been modified to work with fibers.